requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
python_osrsapi==0.0.5
Requests==2.28.2
